all tests configure

+++ src/user-repository.js
@@ -4,7 +4,7 @@ class UserRepository {
     }
 
     async findOneByEmail(email){
-        const user =  await this.collection.findOne({email})
+        const user = await this.collection.findOne({email})
 
         if(user == null){
             throw new Error("User with specified e-mail does not exist")              
@@ -13,7 +13,7 @@ class UserRepository {
     }
 
     async findOneById(id){
-        const user =  await this.collection.findOne({_id: id});
+        const user = await this.collection.findOne({_id: id});
 
         if(user === null){
             throw new Error("User with specified id does not exist")              
@@ -23,11 +23,32 @@ class UserRepository {
     }
 
     async insert(user){
+        const result = await this.collection.findOne({email: user.email});
+        
+        if(result){
+            throw new Error("The provided e-mail is already registered")              
+        }
+
         await this.collection.insertOne(user)
         return user
     }
 
+    async update(userEmail, newUserName){
+        const user =  await this.collection.findOne({email: userEmail});
+        if(user === null){
+            throw new Error("User with specified e-mail does not exist")              
+        }
+        
+        const filter = { email: userEmail }
+        const updateOperation = { $set: { name:newUserName } }
+        await this.collection.updateOne(filter, updateOperation)
+    }
+
     async delete(id){
+        const user = await this.collection.findOne({_id:id});
+        if(user === null){
+            throw new Error("Specified user was not found")              
+        }
         await this.collection.deleteOne({_id: id})
     }
